@@ -1,22 +1,14 @@
-import { Route, Redirect } from "react-router-dom";
-import { existsToken } from "../Helpers/token";
+import { Navigate, Outlet } from "react-router-dom";
+import { isValidToken } from "../Helpers/token";
 import LoaderPage from "./Loaders/LoaderPage";
 import useUserInfo from "./Hooks/useUserInfo";
 
-export default function PrivateRoute({ component: Component, ...rest }) {
+export default function PrivateRoute({ ...props }) {
   const { user, isLoading } = useUserInfo();
 
   if (isLoading) return <LoaderPage />;
 
-  return (
-    <Route {...rest}>
-      {(props) =>
-        existsToken() && user ? (
-          <Component {...props} />
-        ) : (
-          <Redirect to="/" />
-        )
-      }
-    </Route>
-  );
+  if (user && isValidToken()) return <Outlet {...props} />;
+
+  return <Navigate to="/" replace />;
 }
