@@ -1,4 +1,4 @@
-import { Switch, BrowserRouter, Route } from "react-router-dom";
+import { Routes, BrowserRouter, Route } from "react-router-dom";
 import PrivateRoute from "./PrivateRoute";
 import RedirectRoute from "./RedirectRoute";
 import routers from "../config/routers";
@@ -6,14 +6,29 @@ import routers from "../config/routers";
 export default function Routers() {
   return (
     <BrowserRouter>
-      <Switch>
-        {routers.map((route, i) => {
-          const key = route.path || i;
-          if (route.private) return <PrivateRoute {...route} key={key} />;
-          if (route.redirect) return <RedirectRoute {...route} key={key} />;
-          return <Route {...route} key={key} />;
+      <Routes>
+        {routers.map(({ path, element: Element, ...props }, i) => {
+          const key = path || i;
+
+          if (props.private) {
+            return (
+              <Route path={path} element={<PrivateRoute />} key={key}>
+                <Route element={<Element />} index {...props} />
+              </Route>
+            );
+          }
+          if (props.redirect) {
+            return (
+              <Route path={path} element={<RedirectRoute />} key={key}>
+                <Route element={<Element />} index {...props} />
+              </Route>
+            );
+          }
+          return (
+            <Route element={<Element />} path={path} key={key} {...props} />
+          );
         })}
-      </Switch>
+      </Routes>
     </BrowserRouter>
   );
 }
