@@ -1,19 +1,19 @@
 import { useEffect } from "react";
 import { useQuery } from "react-query";
 import { getUserInfo } from "../../Helpers/api";
-import { existsToken } from "../../Helpers/token";
+import { isValidToken } from "../../Helpers/token";
 import useCurrentUser from "./useCurrentUser";
 
 export default function useUserInfo() {
   const { user, setUser, logout } = useCurrentUser();
   const { data, isError, ...args } = useQuery("user", getUserInfo, {
-    enabled: existsToken(),
+    enabled: isValidToken() && !user,
   });
 
   useEffect(() => {
     // It check if there isn't errors, if the token exists, if the user data isn't
     // an empty object and not to override the current user object data
-    if (existsToken()) {
+    if (isValidToken()) {
       if (!isError && data && !user) {
         setUser(data);
       } else if (isError) {
@@ -22,5 +22,5 @@ export default function useUserInfo() {
     }
   }, [data, isError, setUser, logout, user]);
 
-  return { user: data, isError, ...args };
+  return { user: data || user, isError, ...args };
 }
