@@ -1,11 +1,11 @@
-import Spinner from "react-loader-spinner";
+import { Oval } from "react-loader-spinner";
 
 const Loader = () => {
   return (
-    <div className="loader-page">
-      <Spinner
+    <div className="loader-page"> 
+      <Oval
+        ariaLabel="Loading Page"
         visible={true}
-        type="Oval"
         color="rgba(0, 0, 0, 0.388)"
         height={50}
         width={50}
