@@ -1,5 +1,5 @@
-export const api = "https://mern-template-api.vercel.app/api/";
-export const login = "/auth/login";
-export const signup = "/auth/signup";
-export const userInfo = "/user/user";
+export const api = "http://localhost:5000/api/";
+export const login = "auth/login";
+export const signup = "auth/signup";
+export const userInfo = "user";
 export const recaptcha_key= "6Le_BFAUAAAAAIuL8CTLqHOsLC8bto8XS-l3MuTf";
